# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/private/user/save": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new user",
                "operationId": "3612b5fb539111690d361b515bd480b6",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "$ref": "#/components/schemas/UserSchema"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "Vincent",
                                    "email": "vincent@test.com",
                                    "password": "Password@123",
                                    "setting": {
                                        "role_id": 1,
                                        "team_leader_id": 0,
                                        "underwriter_id": 0,
                                        "is_underwriter": 0,
                                        "is_round_robin": 0,
                                        "agent_type": 0,
                                        "renewal_deals": 0,
                                        "insurance_type": 1,
                                        "status": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User save successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Vincent"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "vincent@test.com"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 101
                                                },
                                                "status_text": {
                                                    "type": "string",
                                                    "example": "Active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/private/user/update": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Update user by id",
                "operationId": "99f8ca461f67e9241cf201aeae9eb293",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "": {
                                        "$ref": "#/components/schemas/UserSchema"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 100,
                                    "name": "Vincent",
                                    "email": "vincent@test.com",
                                    "password": "Password@123",
                                    "setting": {
                                        "role_id": 1,
                                        "team_leader_id": 0,
                                        "underwriter_id": 0,
                                        "is_underwriter": 0,
                                        "is_round_robin": 0,
                                        "agent_type": 0,
                                        "renewal_deals": 0,
                                        "insurance_type": 1,
                                        "status": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User save successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Vincent"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "vincent@test.com"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-02-23T16:39:54.000000Z"
                                                },
                                                "id": {
                                                    "type": "number",
                                                    "example": 101
                                                },
                                                "status_text": {
                                                    "type": "string",
                                                    "example": "Active"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "maxItems": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserSchema": {
                "properties": {
                    "": {
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string"
                            },
                            "password": {
                                "type": "string"
                            },
                            "setting": {
                                "properties": {
                                    "role_id": {
                                        "type": "number"
                                    },
                                    "team_leader_id": {
                                        "type": "number"
                                    },
                                    "underwriter_id": {
                                        "type": "number"
                                    },
                                    "is_underwriter": {
                                        "type": "number"
                                    },
                                    "is_round_robin": {
                                        "type": "number"
                                    },
                                    "agent_type": {
                                        "type": "number"
                                    },
                                    "renewal_deals": {
                                        "type": "number"
                                    },
                                    "insurance_type": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}